# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Diet Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserProfileGender =
  (typeof UserProfileGender)[keyof typeof UserProfileGender];

export const UserProfileGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export type UserProfileActivityLevel =
  (typeof UserProfileActivityLevel)[keyof typeof UserProfileActivityLevel];

export const UserProfileActivityLevel = {
  sedentary: "sedentary",
  lightly_active: "lightly_active",
  moderately_active: "moderately_active",
  very_active: "very_active",
  extra_active: "extra_active",
} as const;

export type UserProfileDietPreference =
  (typeof UserProfileDietPreference)[keyof typeof UserProfileDietPreference];

export const UserProfileDietPreference = {
  vegetarian: "vegetarian",
  vegan: "vegan",
  non_vegetarian: "non_vegetarian",
  keto: "keto",
  paleo: "paleo",
  mediterranean: "mediterranean",
  gluten_free: "gluten_free",
} as const;

export type UserProfileGoal =
  (typeof UserProfileGoal)[keyof typeof UserProfileGoal];

export const UserProfileGoal = {
  weight_loss: "weight_loss",
  muscle_gain: "muscle_gain",
  maintenance: "maintenance",
  improve_health: "improve_health",
} as const;

export interface UserProfile {
  id: number;
  userId: number;
  age: number;
  gender: UserProfileGender;
  heightCm: number;
  weightKg: number;
  activityLevel: UserProfileActivityLevel;
  dietPreference: UserProfileDietPreference;
  goal: UserProfileGoal;
  dailyCalorieTarget?: number;
  updatedAt?: string;
}

export type UserProfileBodyGender =
  (typeof UserProfileBodyGender)[keyof typeof UserProfileBodyGender];

export const UserProfileBodyGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export type UserProfileBodyActivityLevel =
  (typeof UserProfileBodyActivityLevel)[keyof typeof UserProfileBodyActivityLevel];

export const UserProfileBodyActivityLevel = {
  sedentary: "sedentary",
  lightly_active: "lightly_active",
  moderately_active: "moderately_active",
  very_active: "very_active",
  extra_active: "extra_active",
} as const;

export type UserProfileBodyDietPreference =
  (typeof UserProfileBodyDietPreference)[keyof typeof UserProfileBodyDietPreference];

export const UserProfileBodyDietPreference = {
  vegetarian: "vegetarian",
  vegan: "vegan",
  non_vegetarian: "non_vegetarian",
  keto: "keto",
  paleo: "paleo",
  mediterranean: "mediterranean",
  gluten_free: "gluten_free",
} as const;

export type UserProfileBodyGoal =
  (typeof UserProfileBodyGoal)[keyof typeof UserProfileBodyGoal];

export const UserProfileBodyGoal = {
  weight_loss: "weight_loss",
  muscle_gain: "muscle_gain",
  maintenance: "maintenance",
  improve_health: "improve_health",
} as const;

export interface UserProfileBody {
  /**
   * @minimum 1
   * @maximum 120
   */
  age: number;
  gender: UserProfileBodyGender;
  /**
   * @minimum 50
   * @maximum 300
   */
  heightCm: number;
  /**
   * @minimum 10
   * @maximum 500
   */
  weightKg: number;
  activityLevel: UserProfileBodyActivityLevel;
  dietPreference: UserProfileBodyDietPreference;
  goal: UserProfileBodyGoal;
}

export interface GeneratePlanBody {
  profileId: number;
}

export interface DietPlan {
  id: number;
  userId: number;
  title: string;
  dailyCalories: number;
  proteinGrams: number;
  carbsGrams: number;
  fatGrams: number;
  /** JSON stringified 7-day meal plan */
  planData: string;
  /** JSON stringified grocery list */
  groceryList?: string;
  createdAt: string;
}

export interface DietPlanSummary {
  id: number;
  title: string;
  dailyCalories: number;
  createdAt: string;
}

export interface CheckFoodBody {
  foodName: string;
  portionSize: string;
}

export interface FoodNutrition {
  foodName: string;
  portionSize: string;
  calories: number;
  proteinGrams: number;
  carbsGrams: number;
  fatGrams: number;
  fiberGrams: number;
  /** 1-10 health score */
  healthScore: number;
  alternatives: string[];
  tips: string;
}

export type LogMealBodyMealType =
  (typeof LogMealBodyMealType)[keyof typeof LogMealBodyMealType];

export const LogMealBodyMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface LogMealBody {
  mealType: LogMealBodyMealType;
  foodName: string;
  portionSize: string;
  calories: number;
  proteinGrams?: number;
  carbsGrams?: number;
  fatGrams?: number;
  loggedAt?: string;
}

export interface MealEntry {
  id: number;
  userId: number;
  mealType: string;
  foodName: string;
  portionSize: string;
  calories: number;
  proteinGrams?: number;
  carbsGrams?: number;
  fatGrams?: number;
  loggedAt: string;
  createdAt: string;
}

export interface LogWaterBody {
  amountMl: number;
  loggedAt?: string;
}

export interface WaterEntry {
  id: number;
  userId: number;
  amountMl: number;
  loggedAt: string;
  createdAt: string;
}

export interface LogWeightBody {
  weightKg: number;
  loggedAt?: string;
}

export interface WeightEntry {
  id: number;
  userId: number;
  weightKg: number;
  loggedAt: string;
  createdAt: string;
}

export interface DailyLog {
  date: string;
  totalCalories: number;
  totalProteinGrams: number;
  totalCarbsGrams: number;
  totalFatGrams: number;
  totalWaterMl: number;
  meals: MealEntry[];
  calorieTarget: number;
  waterTarget: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListMealLogsParams = {
  from?: string;
  to?: string;
};