# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Diet Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}